+++ src/datetimeServer.c
@@ -728,7 +728,7 @@ pthread_t *timeServer()
       goto timeServer_clean_exit;
    }
    pthread_detach(*timeServer_thread);
-   fprintf(stderr,"DATATIMESERVER : %x\n", *timeServer_thread);
+   fprintf(stderr,"DATATIMESERVER : %x\n", (unsigned int)*timeServer_thread);
 
    if(timeServer_thread)
       return timeServer_thread;
